EBP-910 resolve import issues
