--- errors.v1.go
@@ -1,7 +1,0 @@
-package middleware
-
-var ErrAuthorizationTokenEmpty = 20003
-var ErrAuthorizationTokenInvalid = 20004
-
-var AuthorizationTokenEmpty = "authorization token not provided"
-var AuthorizationTokenInvalid = "authorization token invalid"
